# Rewrite Array#each in Ruby