@k0kubun 2024-01-18
Given https://bugs.ruby-lang.org/issues/20182#note-5, https://github.com/ruby/ruby/pull/9533 will take this over.